Add download button in prior notification form
